move common network code to seperate class
