+++ MqttReceiver.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <Arduino.h>
+
+class MqttReceiver
+{
+public:
+    virtual void onMqttDataReceived(char*& topic, byte*& payload, unsigned int& length) = 0;
+};
